Map over multiple inputs simultaneously and return a vector of cognostics data frames

Map over multiple inputs simultaneously and return a vector of cognostics data frames

map2_cog(.x, .y, .f, ...)

pmap_cog(.l, .f, ...)

Arguments

.x, .y
Vectors of the same length. A vector of length 1 will be recycled.

.f
A function, formula, or atomic vector (see map2 for details)

...
additional arguments passed on to .f.

.l
A list of lists. The length of .l determines the number of arguments that .f will be called with. List names will be used if present.

Details

See map2

Examples

# \donttest{
library(tidyr)
library(purrr)
library(rbokeh)
library(dplyr)

iris %>%
  nest(-Species) %>%
  mutate(
    mod = map(data, ~ lm(Sepal.Length ~ Sepal.Width, data = .x)),
    cogs = map2_cog(data, mod, function(data, mod) {
      data_frame(max_sl = max(data$Sepal.Length), slope = coef(mod)[2])
    }),
    panel = map2_plot(data, mod, function(data, mod) {
      figure(xlab = "Sepal.Width", ylab = "Sepal.Length") %>%
        ly_points(data$Sepal.Width, data$Sepal.Length) %>%
        ly_abline(mod)
    })) %>%
  trelliscope(name = "iris")
# }